pib: Fix a race-condition on the send queue
Suppose we had two threads A and B,

1. A enters (%irc-send-thread)
2. A locks irc-send-cv-mutex.
3. A tries to dequeue irc-send-queue and finds that the queue is empty.
4. B calls (irc-send-enqueue!), which means B tries to signal A to
   awake but in fact A is not sleeping *yet*.
5. A unlocks irc-send-cv-mutex and atomically starts to sleep. The
   event that B enqueued gets forgotten in the queue... until someone
   signals A again.

So what was wrong in the first place? We shouldn't have had two
separate mutexes, one for the data structure we wanted to guard (the
send queue in this case), and another for the condition variable. If
our sender and receiver each only had a single mutex, that is, both
irc-recv-cv-mutex and irc-send-cv-mutex had gone:

1. A enters (%irc-send-thread)
2. A locks irc-send-queue-mutex.
3. A tries to dequeue irc-send-queue and finds that the queue is empty.
4. B calls (irc-send-enqueue!) and blocks since irc-send-queue is now
   locked by A. The send queue is still empty here.
5. A unlocks irc-send-queue-mutex and atomically starts to sleep.
6. B awakes, enqueues a new event, and then signals A.
7. A awakes and sees there is an event in the queue.